Feat '*help' function for easier user interaction and add '*killyounglings' and '*good'
Remove urls from images and GIFS.  Update new GIFS to *execute66.
Fix command names for better useability as before was hard to remember.  Feat new
functions '*killyounglings' and '*good' for users to use.

As a result, urls do not appear with the images but upload time is slower.
New features can be implemented to organize actors and movies.